0. Positive anything is better than negative nothing  (automated command)

+++ 0x01-variables_if_else_while/2-print_alphabet.c
@@ -12,7 +12,7 @@
 
 int main(void)
 {
-for (size_t i = 97; i < 123; i++){
+for (int i = 97; i < 123; i++){
 putchar(i);
 }
 putchar('\n');
